feat: add app services: http and contacts
