+++ ClientApp/src/app/components/header/header.component.ts
@@ -5,6 +5,7 @@ import { Component } from '@angular/core';
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss']
 })
+
 export class HeaderComponent {
 
 }
